Add scheduled RSS feed service
