+++ backend/src/main/java/com/backtester/BacktesterApplication.java
@@ -2,7 +2,9 @@ package com.backtester;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
 public class BacktesterApplication {
     public static void main(String[] args) {
