+++ frontend/src/App.jsx
@@ -2,10 +2,12 @@ import React, { useState, useEffect } from 'react'
 import Navbar from './components/Navbar.jsx'
 import Sidebar from './components/Sidebar.jsx'
 import Dashboard from './components/Dashboard.jsx'
+import SentimentTable from './components/SentimentTable.jsx'
 
 export default function App() {
   const [dark, setDark] = useState(false)
   const [sidebar, setSidebar] = useState(false)
+  const [page, setPage] = useState('backtest')
   useEffect(() => {
     document.documentElement.classList.toggle('dark', dark)
   }, [dark])
@@ -28,13 +30,19 @@ export default function App() {
     }
     verify()
   }, [])
+  const navigate = p => {
+    setPage(p)
+    setSidebar(false)
+  }
+
   return (
     <div className="flex h-screen">
-      <Sidebar open={sidebar} onClose={() => setSidebar(false)} />
+      <Sidebar open={sidebar} onClose={() => setSidebar(false)} onNav={navigate} />
       <div className="flex flex-col flex-1">
         <Navbar dark={dark} onToggle={() => setDark(!dark)} onMenu={() => setSidebar(true)} />
         <main className="flex-1 p-4 overflow-y-auto bg-gray-100 dark:bg-gray-900">
-          <Dashboard />
+          {page === 'backtest' && <Dashboard />}
+          {page === 'sentiment' && <SentimentTable />}
         </main>
       </div>
     </div>
